+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- There is now an `options_filter` option for relationship fields that lets you modify the query before getting the relationship options
 - Custom actions and saves now rebuild the supplied config file after performing the action
 - The `editable` property now accepts a closure and is passed the current page's data object
 - New translations (da, it)